Moved callback into separate file
